Change var defs to size_t

+++ 0-bubble_sort.c
@@ -22,7 +22,7 @@ void swap(int *a, int *b)
  */
 void bubble_sort(int *array, size_t size)
 {
-	unsigned long i, j;
+	size_t i, j;
 	bool is_swapped;
 
 	for (i = 0; i < size - 1; i++)
